requirements.txt
# ==============================================
# Robot Framework AI错误分析器 依赖清单
# 推荐Python版本: 3.8 ~ 3.12
# Python 3.13+ 可能存在Pydantic v1兼容问题，可忽略警告不影响核心功能
# ==============================================

# ------------------------------
# 核心必选依赖
# ------------------------------
# 智谱AI官方SDK，固定0.2.2版本（API接口与新版本不兼容）
zai-sdk==0.2.2
# YAML配置文件解析
pyyaml>=6.0
# HTTP请求库（用于网络请求、异常处理）
requests>=2.31.0
# 模板渲染引擎（用于提示词模板、报告模板）
jinja2>=3.1.0

# ------------------------------
# 可选依赖（按需安装）
# ------------------------------
# Robot Framework框架本身（本地运行测试用例时需要，仅做日志分析不需要）
robotframework>=6.0
# 开发/调试依赖
# ipython>=8.0.0       # 交互式调试
# pytest>=7.0.0        # 单元测试
# black>=23.0.0        # 代码格式化

# ------------------------------
# 安装说明
# ------------------------------
# 国内用户建议使用清华源加速安装:
# pip install -r requirements.txt -i https://pypi.tuna.tsinghua.edu.cn/simple
#
# 注意事项:
# 1. 请不要将API Key硬编码到代码中，建议使用环境变量注入
# 2. zai-sdk 0.2.2版本为当前稳定适配版本，请勿随意升级
# 3. Python 3.14运行时出现的Pydantic V1警告可忽略，不影响核心功能
